third_party/blink/web_tests/external/wpt/editing/run/forwarddelete_2001-3000-expected.txt

This is a testharness.js-based test.
Found 14 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["defaultparagraphseparator","div"],["forwarddelete",""]] "foo[]<blockquote><div>bar</div></blockquote>" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["stylewithcss","true"],["forwarddelete",""]] "foo[]<blockquote style=\\"color: blue\\">bar</blockquote>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["forwarddelete",""]] "foo[]<blockquote style=\\"color: blue\\">bar</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font color=\\"#0000ff\\">bar</font>" but got "foo<span style=\\"color:rgb(0, 0, 255)\\">bar</span>"
[FAIL] [["defaultparagraphseparator","div"],["forwarddelete",""]] "foo[]<blockquote><div><p>bar<p>baz</div></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobar<blockquote><div><p>baz</p></div></blockquote>" but got "foobar<blockquote><p>baz</p></blockquote>"
[FAIL] [["defaultparagraphseparator","p"],["forwarddelete",""]] "foo[]<blockquote><div><p>bar<p>baz</div></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobar<blockquote><div><p>baz</p></div></blockquote>" but got "foobar<blockquote><p>baz</p></blockquote>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "foo[]<blockquote style=\\"color: blue\\"><p>bar<p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font color=\\"#0000ff\\">bar</font><blockquote style=\\"color:rgb(0, 0, 255)\\"><p>baz</p></blockquote>" but got "foo<span style=\\"color:rgb(0, 0, 255)\\">bar</span><blockquote style=\\"color:rgb(0, 0, 255)\\"><p>baz</p></blockquote>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "foo[]<blockquote style=\\"color: blue\\"><p>bar<p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font color=\\"#0000ff\\">bar</font><blockquote style=\\"color:rgb(0, 0, 255)\\"><p>baz</p></blockquote>" but got "foo<span style=\\"color:rgb(0, 0, 255)\\">bar</span><blockquote style=\\"color:rgb(0, 0, 255)\\"><p>baz</p></blockquote>"
[FAIL] [["defaultparagraphseparator","div"],["forwarddelete",""]] "foo[]<blockquote><p><span>bar</span><p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span>bar</span><blockquote><p>baz</p></blockquote>" but got "foobar<blockquote><p>baz</p></blockquote>"
[FAIL] [["defaultparagraphseparator","p"],["forwarddelete",""]] "foo[]<blockquote><p><span>bar</span><p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span>bar</span><blockquote><p>baz</p></blockquote>" but got "foobar<blockquote><p>baz</p></blockquote>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<p style=color:blue>foo[]<p>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"color:rgb(0, 0, 255)\\">foo</span>bar</p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foobar</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<p style=color:blue>foo[]<p>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font color=\\"#0000ff\\">foo</font>bar</p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foobar</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<p style=color:blue>foo[]<p>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"color:rgb(0, 0, 255)\\">foo</span>bar</p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foobar</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<p style=color:blue>foo[]<p>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font color=\\"#0000ff\\">foo</font>bar</p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foobar</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<p style=color:blue>foo[]<p style=color:brown>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"color:rgb(0, 0, 255)\\">foo<font color=\\"#a52a2a\\">bar</font></p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span></p>"
Harness: the test ran to completion.
